# assets/openapi/sonos-hub.v1.yaml
openapi: 3.0.3
info:
  title: Sonos Smart Routines Hub API
  version: '1.0.0'
  description: |
    REST API for the Sonos Smart Routines Hub - a LAN-only home automation system for
    Sonos speakers. Provides device discovery, scene/routine management, music playback
    control, and integration with Apple Music. Designed for local-first operation with
    observability via Prometheus metrics and structured JSON logging.
servers:
  - url: http://hub.local:8787
...
paths:
...
  /v1/routine-templates/{template_id}:
    get:
      operationId: getRoutineTemplate
      tags: [templates]
      summary: Get routine template
      description: Get a specific routine template by ID
      parameters:
        - in: path
          name: template_id
          description: Template identifier
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoutineTemplateResponse' }

  # =========================================================================
  # EXECUTIONS ENDPOINTS
  # =========================================================================
...
components:
  schemas:
...
    WeeklySchedule:
      type: object
      required: [type, weekdays, time]
      properties:
        type:
          type: string
          enum: [weekly]
          description: Schedule type discriminator
        weekdays:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          minItems: 1
          description: Days of week to run (0=Sunday, 6=Saturday)
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Time of day in HH:MM format (24-hour)
    AnnualSchedule:
      type: object
      required: [type, month, day, time]
      properties:
        type:
          type: string
          enum: [annual]
          description: Schedule type discriminator
        month:
          type: integer
          minimum: 1
          maximum: 12
          description: Month of year (1-12)
        day:
          type: integer
          minimum: 1
          maximum: 31
          description: Day of month (1-31)
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Time of day in HH:MM format (24-hour)
    Schedule:
      oneOf:
        - $ref: '#/components/schemas/WeeklySchedule'
        - $ref: '#/components/schemas/AnnualSchedule'
      discriminator:
        propertyName: type
        mapping:
          weekly: '#/components/schemas/WeeklySchedule'
          annual: '#/components/schemas/AnnualSchedule'
...
    TemplateSuggestedSpeaker:
      type: object
      required: [room]
      properties:
        room: { type: string }
        volume:
          type: integer
          nullable: true
    TemplateMusicPolicyFixed:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [FIXED] }
        sonos_favorite_id:
          type: string
          nullable: true
    TemplateMusicPolicySet:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [ROTATION, SHUFFLE] }
        set_id:
          type: string
          nullable: true
        no_repeat_window_minutes:
          type: integer
          nullable: true
        fallback_behavior:
          type: string
          nullable: true
    TemplateMusicPolicy:
      oneOf:
        - $ref: '#/components/schemas/TemplateMusicPolicyFixed'
        - $ref: '#/components/schemas/TemplateMusicPolicySet'
    TemplateVisualStyle:
      type: object
      required: [icon, image_name, image_url, gradient_colors, accent_color]
      properties:
        icon: { type: string }
        image_name: { type: string }
        image_url: { type: string }
        gradient_colors:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: string }
        accent_color: { type: string }
    RoutineTemplate:
      type: object
      required:
        [
          template_id,
          name,
          description,
          category,
          schedule,
          timezone,
          suggested_speakers,
          music_policy,
          holiday_behavior,
          arc_tv_policy,
          visual_style
        ]
      properties:
        template_id: { type: string }
        name: { type: string }
        description:
          type: string
          nullable: true
        category: { type: string }
        schedule: { $ref: '#/components/schemas/Schedule' }
        timezone: { type: string }
        suggested_speakers:
          type: array
          items: { $ref: '#/components/schemas/TemplateSuggestedSpeaker' }
        music_policy:
          allOf:
            - $ref: '#/components/schemas/TemplateMusicPolicy'
          nullable: true
        holiday_behavior: { type: string }
        arc_tv_policy: { type: string }
        visual_style: { $ref: '#/components/schemas/TemplateVisualStyle' }
...
    RoutineTemplateResponse:
      type: object
      required: [template]
      properties:
        template: { $ref: '#/components/schemas/RoutineTemplate' }
...